Replace sfslite with MbedTLS
Libbyz uses sfslite for encryption and cryptographic signatures.
However, this library is no longer maintained and does not compile on
modern gcc or clang versions. Furthermore, our project aims at using
libbyz on embedded systems. Therefore, it makes sense to replace sfslite
with a library that is easily available on these systems.  MbedTLS runs
on our target setup (ESP-IDF) as well as on many other embedded
development environments.  Therefore, it is the perfect replacement
choice.

This commit replaces usage of  sfslite code with corresponding calls to
MbedTLS. As a side-effect this commit also replaces the Rabin public key
crypto-system used during session key exchanges with RSA, since the
former is not supported by MbedTLS.

+++ src/Message.h
@@ -10,7 +10,7 @@
 #include "types.h"
 
 // Maximum message size. Must verify ALIGNED_SIZE.
-const int Max_message_size = 9000;
+const size_t Max_message_size = 9000;
 
 //
 // All messages have the following format:
